ethcall is failing

+++ crates/extractor/src/eth_call.rs
@@ -66,12 +66,12 @@ impl EthCall {
         Ok(values)
     }
 
-    pub fn get_chain_list() -> HashMap<String, i32> {
+    pub fn get_chain_list() -> HashMap<String, i64> {
         let file_path = Path::new("chainIds.json");
         let file_content = fs::read_to_string(file_path).expect("Unable to read file");
         let id_to_network: Value = serde_json::from_str(&file_content).expect("Unable to parse JSON");
         
-        let name_to_id: HashMap<String, i32> = id_to_network.as_object().unwrap().iter().map(|(k, v)| (v.as_str().unwrap().to_string(), k.parse::<i32>().unwrap())).collect();
+        let name_to_id: HashMap<String, i64> = id_to_network.as_object().unwrap().iter().map(|(k, v)| (v.as_str().unwrap().to_string(), k.parse::<i64>().unwrap())).collect();
         name_to_id
     }
 }
